libLERC: fix MSVC 2015 Win64 'warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)'
git-svn-id: https://svn.osgeo.org/gdal/trunk@34058 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/mrf/libLERC/Huffman.cpp
@@ -231,7 +231,7 @@ bool Huffman::BuildTreeFromCodes(int& numBitsLUT)
   numBitsLUT = min(maxLen, m_maxNumBitsLUT);
 
   m_decodeLUT.clear();
-  m_decodeLUT.assign(1 << numBitsLUT, pair<short, short>((short)-1, (short)-1));
+  m_decodeLUT.assign(static_cast<size_t>(1 << numBitsLUT), pair<short, short>((short)-1, (short)-1));
 
   for (int i = i0; i < i1; i++)
   {
